+++ examples/mlperf/initializers.py
@@ -34,48 +34,3 @@ class Linear(nn.Linear):
     if bias: self.bias = Tensor.zeros(out_features, dtype=dtypes.float32)
   def __call__(self, x:Tensor):
     return x.linear(self.weight.cast(dtypes.default_float).transpose(), self.bias.cast(dtypes.default_float) if self.bias is not None else None)
-
-class UnsyncedBatchNorm:
-  def __init__(self, sz:int, eps=1e-5, affine=True, track_running_stats=True, momentum=0.1, num_devices=1):
-    self.eps, self.track_running_stats, self.momentum = eps, track_running_stats, momentum
-    self.num_devices = num_devices
-
-    if affine: self.weight, self.bias = Tensor.ones(sz, dtype=dtypes.float32), Tensor.zeros(sz, dtype=dtypes.float32)
-    else: self.weight, self.bias = None, None
-
-    self.running_mean = Tensor.zeros(num_devices, sz, dtype=dtypes.float32, requires_grad=False)
-    self.running_var = Tensor.ones(num_devices, sz, dtype=dtypes.float32, requires_grad=False)
-    self.num_batches_tracked = Tensor.zeros(1, dtype=dtypes.int, requires_grad=False)
-
-  def __call__(self, x:Tensor):
-    if isinstance(x.lazydata, MultiLazyBuffer): assert x.lazydata.axis is None or x.lazydata.axis == 0 and len(x.lazydata.lbs) == self.num_devices
-
-    xr = x.reshape(self.num_devices, -1, *x.shape[1:]).cast(dtypes.float32)
-    batch_mean, batch_invstd = self.calc_stats(xr)
-    ret = xr.batchnorm(
-      self.weight.reshape(1, -1).expand((self.num_devices, -1)),
-      self.bias.reshape(1, -1).expand((self.num_devices, -1)),
-      batch_mean, batch_invstd, axis=(0, 2))
-    return ret.reshape(x.shape).cast(x.dtype)
-
-  def calc_stats(self, x:Tensor):
-    if Tensor.training:
-      # This requires two full memory accesses to x
-      # https://github.com/pytorch/pytorch/blob/c618dc13d2aa23625cb0d7ada694137532a4fa33/aten/src/ATen/native/cuda/Normalization.cuh
-      # There's "online" algorithms that fix this, like https://en.wikipedia.org/wiki/Algorithms_for_calculating_variance#Welford's_Online_algorithm
-      batch_mean = x.mean(axis=(1,3,4))
-      y = (x - batch_mean.reshape(shape=[batch_mean.shape[0], 1, batch_mean.shape[1], 1, 1]))
-      batch_var = (y*y).mean(axis=(1,3,4))
-      batch_invstd = batch_var.add(self.eps).pow(-0.5)
-
-      # NOTE: wow, this is done all throughout training in most PyTorch models
-      if self.track_running_stats:
-        self.running_mean.assign((1-self.momentum) * self.running_mean + self.momentum * batch_mean.detach().cast(self.running_mean.dtype))
-        batch_var_adjust = prod(y.shape[1:]) / (prod(y.shape[1:]) - y.shape[2])
-        self.running_var.assign((1-self.momentum) * self.running_var + self.momentum * batch_var_adjust * batch_var.detach().cast(self.running_var.dtype))
-        self.num_batches_tracked += 1
-    else:
-      batch_mean = self.running_mean
-      # NOTE: this can be precomputed for static inference. we expand it here so it fuses
-      batch_invstd = self.running_var.reshape(self.running_var.shape[0], 1, self.running_var.shape[1], 1, 1).expand(x.shape).add(self.eps).rsqrt()
-    return batch_mean, batch_invstd
